+++ src/main/java/net/lukkit/lukkit/command/completer/BaseCompleter.java
@@ -1,4 +1,4 @@
-package net.lukkit.lukkit.completer;
+package net.lukkit.lukkit.command.completer;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -6,17 +6,16 @@ import org.bukkit.command.TabCompleter;
 
 import java.util.List;
 
-public class DevCompleter implements TabCompleter {
+public class BaseCompleter implements TabCompleter {
     private static final String[] commands = {
-            "pack",
-            "unpack",
-            "compress",
-            "kill",
+            "help",
+            "report",
+            "dev",
+            "plugin"
     };
 
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
-        // "/lukkit plugin list"
         return null;
     }
 }
