+++ src/main/java/net/lukkit/lukkit/command/completer/DevCompleter.java
@@ -1,4 +1,4 @@
-package net.lukkit.lukkit.completer;
+package net.lukkit.lukkit.command.completer;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -6,16 +6,17 @@ import org.bukkit.command.TabCompleter;
 
 import java.util.List;
 
-public class BaseCompleter implements TabCompleter {
+public class DevCompleter implements TabCompleter {
     private static final String[] commands = {
-            "help",
-            "report",
-            "dev",
-            "plugin"
+            "pack",
+            "unpack",
+            "compress",
+            "kill",
     };
 
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
+        // "/lukkit plugin list"
         return null;
     }
 }
