--- src/main/java/net/lukkit/lukkit/packager/Compressor.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.packager;
-
-public class Compressor {
-}
