--- src/main/java/net/lukkit/lukkit/packager/Decompressor.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.packager;
-
-public class Decompressor {
-}
