--- src/main/java/net/lukkit/lukkit/packager/PluginVerifier.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.packager;
-
-public class PluginVerifier {
-}
