--- src/main/java/net/lukkit/lukkit/sandbox/command/LukkitCommand.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.sandbox.command;
-
-public class LukkitCommand {
-}
