--- src/main/java/net/lukkit/lukkit/sandbox/command/LukkitTabComleter.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.sandbox.command;
-
-public class LukkitTabComleter {
-}
