+++ src/main/java/net/lukkit/lukkit/sandbox/plugin/command/LukkitCommand.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.sandbox.plugin.command;
+
+public class LukkitCommand {
+}
