+++ src/main/java/net/lukkit/lukkit/sandbox/plugin/command/LukkitTabComleter.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.sandbox.plugin.command;
+
+public class LukkitTabComleter {
+}
