+++ src/main/java/net/lukkit/lukkit/util/packager/PluginVerifier.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.util.packager;
+
+public class PluginVerifier {
+}
