+++ src/main/java/net/lukkit/lukkit/util/wizard/ChatHandler.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.util.wizard;
+
+public class ChatHandler {
+}
