+++ src/main/java/net/lukkit/lukkit/util/wizard/Wizard.java
@@ -1,0 +1,4 @@
+package net.lukkit.lukkit.util.wizard;
+
+public class Wizard {
+}
