--- src/main/java/net/lukkit/lukkit/wizard/ChatHandler.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.wizard;
-
-public class ChatHandler {
-}
