--- src/main/java/net/lukkit/lukkit/wizard/Wizard.java
@@ -1,4 +1,0 @@
-package net.lukkit.lukkit.wizard;
-
-public class Wizard {
-}
